+++ Part2/courseinfo/src/App.js
@@ -1,54 +1,5 @@
 import React from 'react';
-
-const Header = ({course}) => {
-  //console.log(course)
-  return (
-    <div>
-      <h2>
-        {course.name}
-      </h2>
-    </div>
-  )
-}
-
-const Part = ({part, exercises}) => {
-  //console.log(id);
-  return(
-    <p>
-      {part} {exercises}
-    </p>
-  )
-}
-
-const Content = ({course}) => {
-  return (
-    <div>
-      {course.parts.map(part => 
-        <Part key={part.id} part={part.name} exercises={part.exercises}/>
-      )}
-    </div>
-  )
-}
-
-const Total = ({course}) => {
-  let sumArr = course.parts.map(exercise => exercise.exercises)
-  let sum = sumArr.reduce(function(a, b){return a + b});
-  return (
-    <div>
-      <p><strong>Number of exercises {sum}</strong></p> 
-    </div>
-  )
-}
-
-const Course = ({course}) => {
-  return(
-    <>
-    <Header course={course}/>
-    <Content course={course}/>
-    <Total course={course}/>
-    </>
-  )
-}
+import Course from './components/Course'
 
 const App = () => {
   const courses = [
